Rewrite FloatingToc as pure Astro component
Replaces the React implementation with server-rendered Astro components
and a vanilla JS scroll listener, eliminating the React runtime dependency
for the TOC. Handles initial hash-based anchor navigation on page load.
