+++ apps/site/src/components/floating-toc/types.ts
@@ -1,0 +1,6 @@
+export interface TocEntry {
+  slug: string
+  text: string
+  depth: number
+  items: TocEntry[]
+}
